starting migration to smarter file management
